Delete user bug fixed

+++ src/domains/user/interfaces/impl/repository/user_repo.go
@@ -177,7 +177,10 @@ func (userRepo *userRepository) DestroyUserById(ctx context.Context, Id *primiti
 	var shelter *Shelter.Shelter
 	err = userRepo.database.Collection(collections.ShelterCollectionName).FindOne(ctx, bson.M{"user_id": Id}).Decode(&shelter)
 	if err != nil {
-		return nil, err
+		if err != mongo.ErrNoDocuments {
+			return nil, err
+		}
+		shelter = nil
 	}
 	// Start a session
 	session, err := userRepo.database.Client().StartSession()
@@ -187,31 +190,33 @@ func (userRepo *userRepository) DestroyUserById(ctx context.Context, Id *primiti
 	defer session.EndSession(ctx)
 	// Define the callback function for the transaction
 	callback := func(sessCtx mongo.SessionContext) (interface{}, error) {
-		// Delete from request collection
-		_, errCallback := userRepo.database.Collection(collections.PetCollectionName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
-		if errCallback != nil {
-			return nil, errCallback
-		}
+		if shelter != nil {
+			// Delete from request collection
+			_, errCallback := userRepo.database.Collection(collections.PetCollectionName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
+			if errCallback != nil {
+				return nil, errCallback
+			}
 
-		// Delete from pet collection
-		_, errCallback = userRepo.database.Collection(collections.RequestName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
-		if errCallback != nil {
-			return nil, errCallback
-		}
+			// Delete from pet collection
+			_, errCallback = userRepo.database.Collection(collections.RequestName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
+			if errCallback != nil {
+				return nil, errCallback
+			}
 
-		// Delete from shelter favorite collection
-		_, errCallback = userRepo.database.Collection(collections.ShelterFavoriteName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
-		if errCallback != nil {
-			return nil, errCallback
-		}
+			// Delete from shelter favorite collection
+			_, errCallback = userRepo.database.Collection(collections.ShelterFavoriteName).DeleteMany(sessCtx, bson.M{"shelter_id": shelter.ID})
+			if errCallback != nil {
+				return nil, errCallback
+			}
 
-		// Delete from shelter collection
-		_, errCallback = userRepo.database.Collection(collections.ShelterCollectionName).DeleteMany(sessCtx, bson.M{"_id": shelter.ID})
-		if errCallback != nil {
-			return nil, errCallback
+			// Delete from shelter collection
+			_, errCallback = userRepo.database.Collection(collections.ShelterCollectionName).DeleteMany(sessCtx, bson.M{"_id": shelter.ID})
+			if errCallback != nil {
+				return nil, errCallback
+			}
 		}
 		// Delete from user collection
-		_, errCallback = userRepo.database.Collection(collections.ShelterCollectionName).DeleteMany(sessCtx, bson.M{"_id": Id})
+		_, errCallback := userRepo.database.Collection(collections.UserCollectionName).DeleteMany(sessCtx, bson.M{"_id": Id})
 		if errCallback != nil {
 			return nil, errCallback
 		}
